# server/src/main/java/cz/plc/prx/docker/dash/api/EnvironmentApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package cz.plc.prx.docker.dash.api;

import cz.plc.prx.docker.dash.model.Environment;

import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-10-10T07:43:14.863Z")

@Api(value = "environment", description = "the environment API")
public interface EnvironmentApi {

    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/environment/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> environmentDelete(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "x", notes = "x", response = Environment.class, responseContainer = "List", tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "x", response = Environment.class, responseContainer = "List") })
    
    @RequestMapping(value = "/environment",
        method = RequestMethod.GET)
    ResponseEntity<List<Environment>> environmentGet();


    @ApiOperation(value = "", notes = "", response = Environment.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Environment.class),
        @ApiResponse(code = 404, message = "A container with the specified ID was not found.", response = Void.class) })
    
    @RequestMapping(value = "/environment/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<Environment> environmentGet(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/environment/{id}/restart",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> environmentRestart(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/environment/{id}/start",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> environmentStart(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/environment/{id}/stop",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> environmentStop(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);

}