# server/src/main/java/cz/plc/prx/docker/dash/api/InstanceApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.2.3).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package cz.plc.prx.docker.dash.api;

import cz.plc.prx.docker.dash.model.InstanceExt;

import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2017-10-10T07:43:14.863Z")

@Api(value = "instance", description = "the instance API")
public interface InstanceApi {

    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/instance/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.DELETE)
    ResponseEntity<Void> instanceDelete(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = InstanceExt.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = InstanceExt.class),
        @ApiResponse(code = 404, message = "A container with the specified ID was not found.", response = Void.class) })
    
    @RequestMapping(value = "/instance/{id}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<InstanceExt> instanceGet(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/instance/{id}/restart",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> instanceRestart(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/instance/{id}/start",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> instanceStart(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);


    @ApiOperation(value = "", notes = "", response = Void.class, tags={ "container", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "successful operation", response = Void.class) })
    
    @RequestMapping(value = "/instance/{id}/stop",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> instanceStop(@ApiParam(value = "name of selected container",required=true ) @PathVariable("id") String id);

}